Update `message-create!` to return `ex-info` on failure
This should ideally not be a breaking change because the user cannot rely on
what the returned value would be in a failure case, as they cannot be
distinguished from success easily. As a result, this should simply be an
expansion of current features by making a useful return value in the failure
case, rather than a breaking change.

It was suggested to me that this could be a thrown exception, however this would
be a break from the pattern in discljord where all failure cases are shown in
the return values and are potentially logged.