Fix edit command endpoint implementations
The implementations of the `edit-guild-application-command!` and
`edit-global-application-command!` endpoints were erroneous, because
they both expected the parameters `name` and `description` as mandatory
parameters when in fact they are optional. This resulted in incorrect
behaviour.

Additionally, the way parameters to command creations and edits are
parsed has been improved to allow for flexibility (i.e. allowing users
to use additional parameters that may be added in the future even if
discljord hasn't implemented explicit support for them yet).